refactor: decompose window_event RedrawRequested arm into modules
Breaks up the ~1,160-line `RedrawRequested` arm in `main.rs` into focused
`impl App` methods across binary-local modules, completing the decomposition
deferred from PR #138. Pure relocation — no behaviour change.

- gpu_orchestration: step_simulation, dispatch_sim_ticks, collect_state_readback
- snapshot: record_overlays_and_histories, rebuild_agent_snapshots,
  update_evo_snapshot, update_world_snapshot, build_biome_image
- render_pipeline: rebuild_dynamic_meshes, update_agent_instances,
  rebuild_hud_text, render_frame
- replay_coord (new): record_replay_tick, advance_replay_playback
- ui_chrome (new): draw_top_bar, draw_console, draw_agent_sidebar

main.rs drops from 1,751 to 654 lines; window_event is now a readable
orchestration sequence. All 122 sandbox tests pass.

Closes #113
