+++ crates/xagent-sandbox/src/main.rs
@@ -2,13 +2,14 @@ mod app;
 mod evolution;
 mod gpu_orchestration;
 mod render_pipeline;
+mod replay_coord;
 mod snapshot;
+mod ui_chrome;
 
 use std::sync::Arc;
 use std::time::Instant;
 
 use clap::Parser;
-use glam::Vec3;
 use log::info;
 
 use winit::application::ApplicationHandler;
@@ -18,26 +19,14 @@ use winit::keyboard::{KeyCode, PhysicalKey};
 use winit::window::{WindowAttributes, WindowId};
 use xagent_shared::{BrainConfig, FullConfig, GovernorConfig, WorldConfig};
 
-use xagent_brain::buffers::{
-    FOOD_STATE_STRIDE, F_POS_X, F_POS_Z, F_RESPAWN_TIMER, PHYS_STRIDE, P_ALIVE, P_DEATH_COUNT,
-    P_ENERGY, P_EXPLORATION_RATE_OUT, P_FACING_X, P_FACING_Y, P_FACING_Z, P_FATIGUE_FACTOR_OUT,
-    P_FOOD_COUNT, P_GRADIENT_OUT, P_INTEGRITY, P_LAST_DEATH_TICK, P_MAX_ENERGY, P_MAX_INTEGRITY,
-    P_MOTOR_FWD_OUT, P_MOTOR_TURN_OUT, P_POS_X, P_POS_Y, P_POS_Z, P_PREDICTION_ERROR,
-    P_TICKS_ALIVE, P_URGENCY_OUT, P_VEL_X, P_VEL_Y, P_VEL_Z, P_YAW,
-};
 use xagent_sandbox::agent::{mutate_config, MAX_AGENTS};
 use xagent_sandbox::headless;
-use xagent_sandbox::overlay;
-use xagent_sandbox::renderer::font::TextItem;
 use xagent_sandbox::renderer::{GpuMesh, InstanceData, Renderer};
-use xagent_sandbox::ui::{
-    AgentSnapshot, EguiIntegration, EvolutionAction, EvolutionState, SortMode, Tab, TabContext,
-};
+use xagent_sandbox::ui::{EguiIntegration, EvolutionAction, EvolutionState};
 
 use xagent_sandbox::world::WorldState;
 
-use crate::app::{App, REBUILD_THROTTLE, SIM_DT, SIM_RATE};
-use crate::snapshot::record_agent_histories;
+use crate::app::{App, SIM_RATE};
 
 // ── CLI ────────────────────────────────────────────────────────────────
 
@@ -498,1119 +487,33 @@ impl ApplicationHandler for App {
                 }
 
                 // Ensure background kernel creation is started if needed and
-                // collected when ready, even while the tick gate below
-                // suppresses accumulation.
+                // collected when ready.
                 self.ensure_gpu_kernel();
 
-                // ── simulation ticks (fixed timestep, adaptive budget) ──
-                // Skip ticks while a generation transition is in flight to
-                // avoid GPU contention with the async readback/reset.
-                // Also skip accumulation when the kernel is being recreated
-                // in the background to prevent catch-up hitch when it lands.
-                if !self.paused && self.gen_transition.is_none() && self.pending_kernel.is_none() {
-                    let sim_delta_time = SIM_DT as f64;
-                    self.sim_accumulator += dt as f64 * self.speed_multiplier as f64;
-                    // Only dispatch when the accumulator has enough for at
-                    // least brain_tick_stride ticks, so every dispatch
-                    // includes a brain cycle and produces motor commands.
-                    // Sub-stride dispatches would be physics-only (no brain
-                    // cycles), leaving agents with stale motor outputs.
-                    // The accumulator keeps its fractional remainder across
-                    // frames so no sim-time is lost.
-                    let min_dispatch = self
-                        .gpu_kernel
-                        .as_ref()
-                        .map_or(10, |kernel| kernel.brain_tick_stride());
-                    // Cap must allow at least min_dispatch ticks to
-                    // accumulate, otherwise low speeds (1x, 2x) can never
-                    // reach the dispatch threshold.
-                    let max_accumulator = sim_delta_time
-                        * (self.speed_multiplier as f64 * 3.0).max(min_dispatch as f64 + 2.0);
-                    self.sim_accumulator = self.sim_accumulator.min(max_accumulator);
-                    let raw_ticks = ((self.sim_accumulator / sim_delta_time) as u32)
-                        .min(self.gpu_tick_budget)
-                        .min(500);
-                    let ticks_to_run = if raw_ticks >= min_dispatch {
-                        raw_ticks
-                    } else {
-                        0
-                    };
-
-                    if ticks_to_run > 0 {
-                        if let Some(ref mut kernel) = self.gpu_kernel {
-                            kernel.dispatch_batch(self.tick, ticks_to_run);
-
-                            self.sim_accumulator -= ticks_to_run as f64 * sim_delta_time;
-
-                            self.gpu_tick_budget =
-                                (self.gpu_tick_budget + self.gpu_tick_budget / 4 + 1).min(64_000);
-
-                            self.tick += ticks_to_run as u64;
-                            self.tps_tick_count += ticks_to_run as u64;
-                            self.snap_dirty = true;
-
-                            if let Some(gov) = &mut self.governor {
-                                for _ in 0..ticks_to_run {
-                                    gov.tick();
-                                }
-                            }
-
-                            // Async telemetry readback for the selected agent.
-                            // `request_agent_telemetry` is gated internally: it no-ops
-                            // if a readback for the same agent is already pending, and
-                            // clears the old pending if the agent changed.
-                            if self.selected_agent_idx < self.agents.len() {
-                                let brain_idx = self.agents[self.selected_agent_idx].brain_idx;
-
-                                kernel.request_agent_telemetry(brain_idx);
-
-                                // Collect any completed readback (non-blocking)
-                                if let Some(tel) = kernel.try_collect_telemetry() {
-                                    let a = &mut self.agents[self.selected_agent_idx];
-                                    // Only update fields NOT already populated by
-                                    // the every-frame physics readback (below).
-                                    // Physics readback sets: cached_motor,
-                                    // cached_gradient, cached_urgency,
-                                    // cached_fatigue_factor, cached_prediction_error,
-                                    // cached_exploration_rate.
-                                    a.cached_frame.vision.color = tel.vision_color;
-                                    a.cached_mean_attenuation = tel.mean_attenuation;
-                                    a.cached_curiosity_bonus = tel.curiosity_bonus;
-                                    a.cached_staleness = tel.staleness;
-                                }
-                            }
-
-                            self.food_dirty = true;
-
-                            // Record tick for replay using the latest async physics
-                            // readback plus per-agent cached telemetry. Position/yaw/
-                            // alive/energy/integrity come from the physics readback
-                            // stored on each agent body. Motor outputs, gradient/
-                            // urgency, and the exploration/prediction/attenuation/
-                            // curiosity/fatigue/motor-variance fields come from
-                            // `cached_*` telemetry; in this path only the selected
-                            // agent's cache is refreshed each frame.
-                            // `credit_magnitude`, `patterns_recalled`, `phase`, and
-                            // `vision_color` are not populated here and are left as
-                            // defaults. `raw_gradient` mirrors `gradient` because
-                            // there is no separate raw gradient source in this GPU
-                            // readback path.
-                            if let Some(ref mut rec) = self.recording {
-                                // Async readback snapshots arrive independently of
-                                // governor tick advancement, so record them at the
-                                // next dense replay index.
-                                let tick = rec.total_ticks;
-                                let records: Vec<xagent_sandbox::replay::TickRecord> = self.agents.iter().map(|a| {
-                                    xagent_sandbox::replay::TickRecord {
-                                        position: [a.body.body.position.x, a.body.body.position.y, a.body.body.position.z],
-                                        yaw: a.body.yaw,
-                                        alive: a.body.body.alive,
-                                        energy: a.body.body.internal.energy,
-                                        integrity: a.body.body.internal.integrity,
-                                        motor_forward: a.cached_motor.forward,
-                                        motor_turn: a.cached_motor.turn,
-                                        exploration_rate: a.cached_exploration_rate,
-                                        prediction_error: a.cached_prediction_error,
-                                        gradient: a.cached_gradient,
-                                        raw_gradient: a.cached_gradient,
-                                        urgency: a.cached_urgency,
-                                        credit_magnitude: 0.0,
-                                        patterns_recalled: 0,
-                                        phase: xagent_sandbox::replay::GenerationRecording::phase_to_u8("RANDOM"),
-                                        mean_attenuation: a.cached_mean_attenuation,
-                                        curiosity_bonus: a.cached_curiosity_bonus,
-                                        fatigue_factor: a.cached_fatigue_factor,
-                                        staleness: a.cached_staleness,
-                                        vision_color: None,
-                                    }
-                                }).collect();
-                                rec.record_tick(tick, &records);
-                            }
-                        }
-                    }
-
-                    // ── Every-frame state readback ──
-                    // Collect GPU staging data and update agent positions
-                    // regardless of whether ticks were dispatched this frame.
-                    // This ensures smooth visual updates at low speed multipliers
-                    // where dispatches happen infrequently.
-                    if let Some(ref mut kernel) = self.gpu_kernel {
-                        if kernel.try_collect_state() {
-                            let state = kernel.cached_state();
-                            for i in 0..self.agents.len() {
-                                let base = i * PHYS_STRIDE;
-                                // Guard every physics field we read below. Use the
-                                // stride's highest offset (PHYS_STRIDE - 1) so adding
-                                // new fields doesn't silently leave the guard stale.
-                                if base + PHYS_STRIDE > state.len() {
-                                    break;
-                                }
-                                let a = &mut self.agents[i];
-                                a.body.body.position = Vec3::new(
-                                    state[base + P_POS_X],
-                                    state[base + P_POS_Y],
-                                    state[base + P_POS_Z],
-                                );
-                                a.body.body.alive = state[base + P_ALIVE] > 0.5;
-                                a.body.yaw = state[base + P_YAW];
-                                a.body.body.internal.energy = state[base + P_ENERGY];
-                                a.body.body.internal.integrity = state[base + P_INTEGRITY];
-                                a.body.body.internal.max_energy = state[base + P_MAX_ENERGY];
-                                a.body.body.internal.max_integrity = state[base + P_MAX_INTEGRITY];
-                                a.body.body.velocity = Vec3::new(
-                                    state[base + P_VEL_X],
-                                    state[base + P_VEL_Y],
-                                    state[base + P_VEL_Z],
-                                );
-                                a.food_consumed = state[base + P_FOOD_COUNT] as u32;
-                                a.total_ticks_alive = state[base + P_TICKS_ALIVE] as u64;
-                                let new_deaths = state[base + P_DEATH_COUNT] as u32;
-                                let gpu_death_tick = state[base + P_LAST_DEATH_TICK] as u64;
-                                a.apply_death_count_readback(
-                                    new_deaths,
-                                    gpu_death_tick,
-                                    self.tick.saturating_sub(1),
-                                );
-                                a.body.body.facing = Vec3::new(
-                                    state[base + P_FACING_X],
-                                    state[base + P_FACING_Y],
-                                    state[base + P_FACING_Z],
-                                );
-                                a.cached_prediction_error = state[base + P_PREDICTION_ERROR];
-                                a.cached_exploration_rate = state[base + P_EXPLORATION_RATE_OUT];
-                                a.cached_fatigue_factor = state[base + P_FATIGUE_FACTOR_OUT];
-                                a.cached_motor.forward = state[base + P_MOTOR_FWD_OUT];
-                                a.cached_motor.turn = state[base + P_MOTOR_TURN_OUT];
-                                a.cached_gradient = state[base + P_GRADIENT_OUT];
-                                a.cached_urgency = state[base + P_URGENCY_OUT];
-                            }
-
-                            self.snap_dirty = true;
-                            self.hud_dirty = true;
-
-                            // Diagnostic: log first readback Y vs terrain height
-                            if !self.readback_logged {
-                                if let Some(world) = &self.world {
-                                    let n = self.agents.len().min(5);
-                                    if n > 0 {
-                                        for i in 0..n {
-                                            let a = &self.agents[i];
-                                            let p = a.body.body.position;
-                                            let terrain_y = world.terrain.height_at(p.x, p.z);
-                                            log::info!(
-                                                "[TERRAIN-DIAG] Agent {} pos=({:.2}, {:.2}, {:.2}) terrain_y={:.2} diff={:.2}",
-                                                i, p.x, p.y, p.z, terrain_y, p.y - terrain_y,
-                                            );
-                                        }
-                                        self.readback_logged = true;
-                                    }
-                                }
-                            }
-                        }
-                    }
-
-                    // Heatmap + trail recording
-                    if let Some(world) = &self.world {
-                        for agent in &mut self.agents {
-                            if agent.body.body.alive {
-                                agent.record_heatmap(world.config.world_size);
-                                agent.record_trail();
-                            }
-                        }
-                    }
-
-                    // Sparkline histories
-                    for agent in &mut self.agents {
-                        record_agent_histories(agent);
-                    }
-
-                    self.heatmap_dirty = true;
-
-                    // ── Generation completion check (after tick batch) ──
-                    if self.gen_transition.is_none() {
-                        if let Some(gov) = &self.governor {
-                            if gov.generation_complete() {
-                                self.advance_generation();
-                            }
-                        }
-                    }
-                }
+                // ── simulation ticks (fixed timestep) + every-frame readback ──
+                self.step_simulation(dt);
 
                 // Drive the multi-frame generation transition state machine
                 // (runs even when sim ticks are paused/skipped due to transition).
                 self.poll_gen_transition();
 
                 // Advance replay playback
-                if self.replay_state.active && self.replay_state.playing {
-                    let advance = (self.replay_state.speed as u64).max(1);
-                    self.replay_state.current_tick = (self.replay_state.current_tick + advance)
-                        .min(self.replay_state.total_ticks.saturating_sub(1));
-                    if self.replay_state.current_tick
-                        >= self.replay_state.total_ticks.saturating_sub(1)
-                    {
-                        self.replay_state.playing = false;
-                    }
-                }
+                self.advance_replay_playback();
 
                 // Fix selected index if agents were removed
                 if !self.agents.is_empty() {
                     self.selected_agent_idx = self.selected_agent_idx.min(self.agents.len() - 1);
                 }
 
-                // ── rebuild dynamic meshes (throttled to ~10 Hz) ──────
-                let mesh_rebuild_due =
-                    self.paused || self.last_mesh_rebuild.elapsed() >= REBUILD_THROTTLE;
+                // ── rebuild dynamic meshes (throttled) + per-frame marker ──
+                self.rebuild_dynamic_meshes();
 
-                let mut did_rebuild = false;
+                // ── update agent instance buffer + HUD/text (only when dirty) ──
+                self.update_agent_instances();
+                self.rebuild_hud_text();
 
-                if self.food_dirty && mesh_rebuild_due {
-                    if let (Some(renderer), Some(world), Some(food_gpu)) =
-                        (&self.renderer, &self.world, &mut self.food_gpu)
-                    {
-                        let fm = world.food_mesh();
-                        food_gpu.update_from_mesh(&renderer.queue, &fm);
-                        self.food_dirty = false;
-                        did_rebuild = true;
-                    }
-                }
-
-                // ── rebuild heatmap overlay ─────────────────────────
-                if self.heatmap_enabled && self.heatmap_dirty {
-                    if let (Some(renderer), Some(world), Some(heatmap_gpu)) =
-                        (&self.renderer, &self.world, &mut self.heatmap_gpu)
-                    {
-                        if let Some(agent) = self.agents.get(self.selected_agent_idx) {
-                            let mesh = overlay::build_heatmap_mesh(
-                                &agent.heatmap,
-                                world.config.world_size,
-                                &world.terrain,
-                            );
-                            heatmap_gpu.update_from_mesh(&renderer.queue, &mesh);
-                        }
-                    }
-                    self.heatmap_dirty = false;
-                } else if !self.heatmap_enabled {
-                    if let Some(heatmap_gpu) = &mut self.heatmap_gpu {
-                        heatmap_gpu.num_indices = 0;
-                    }
-                }
-
-                // ── rebuild trail overlay for ALL agents (throttled) ──
-                if mesh_rebuild_due {
-                    if let (Some(renderer), Some(trail_gpu)) = (&self.renderer, &mut self.trail_gpu)
-                    {
-                        let any_dirty = self.agents.iter().any(|a| a.trail_dirty);
-                        if any_dirty {
-                            let agent_data: Vec<(&[[f32; 3]], &[f32; 3], bool)> = self
-                                .agents
-                                .iter()
-                                .map(|a| {
-                                    (a.trail.as_slice(), &a.color as &[f32; 3], a.body.body.alive)
-                                })
-                                .collect();
-                            let mesh = overlay::build_all_trails_mesh(&agent_data);
-                            if mesh.indices.is_empty() {
-                                trail_gpu.num_indices = 0;
-                            } else {
-                                trail_gpu.update_from_mesh(&renderer.queue, &mesh);
-                            }
-                            for a in &mut self.agents {
-                                a.trail_dirty = false;
-                            }
-                            did_rebuild = true;
-                        }
-                    }
-                }
-
-                if did_rebuild && !self.paused {
-                    self.last_mesh_rebuild = Instant::now();
-                }
-
-                // ── rebuild selection marker above focused agent ──────
-                if let (Some(renderer), Some(marker_gpu)) = (&self.renderer, &mut self.marker_gpu) {
-                    if let Some(agent) = self.agents.get(self.selected_agent_idx) {
-                        if agent.body.body.alive {
-                            let mesh = overlay::build_marker_mesh(agent.body.body.position);
-                            marker_gpu.update_from_mesh(&renderer.queue, &mesh);
-                        } else {
-                            marker_gpu.num_indices = 0;
-                        }
-                    } else {
-                        marker_gpu.num_indices = 0;
-                    }
-                }
-
-                // ── update agent instance buffer (only when sim ticked) ──
-                if self.hud_dirty {
-                    let instances: Vec<InstanceData> = self
-                        .agents
-                        .iter()
-                        .map(|a| {
-                            let color = if !a.body.body.alive {
-                                use xagent_sandbox::agent::srgb_to_linear;
-                                [
-                                    srgb_to_linear(0.25),
-                                    srgb_to_linear(0.25),
-                                    srgb_to_linear(0.25),
-                                ]
-                            } else {
-                                // Use the agent's assigned palette color, converted
-                                // from sRGB to linear so the sRGB framebuffer
-                                // produces the correct final color matching the sidebar.
-                                use xagent_sandbox::agent::srgb_to_linear;
-                                [
-                                    srgb_to_linear(a.color[0]),
-                                    srgb_to_linear(a.color[1]),
-                                    srgb_to_linear(a.color[2]),
-                                ]
-                            };
-                            InstanceData {
-                                position: a.body.body.position.into(),
-                                color,
-                                scale: 2.0,
-                                _pad: 0.0,
-                            }
-                        })
-                        .collect();
-                    self.agent_instance_count = instances.len() as u32;
-
-                    if let (Some(renderer), Some(buf)) =
-                        (&self.renderer, &self.agent_instance_buffer)
-                    {
-                        if !instances.is_empty() {
-                            renderer
-                                .queue
-                                .write_buffer(buf, 0, bytemuck::cast_slice(&instances));
-                        }
-                    }
-                }
-
-                // ── update HUD + text geometry (only when dirty) ──
-                if self.hud_dirty {
-                    self.cached_hud_bars = self.build_hud_bars();
-
-                    let mut text_items: Vec<TextItem> = Vec::new();
-
-                    // FPS counter at top-right
-                    text_items.push(TextItem {
-                        text: format!("FPS: {:.0}", self.fps),
-                        x: 0.62,
-                        y: 0.97,
-                        scale: 0.04,
-                        color: [0.0, 0.0, 0.0, 1.0],
-                    });
-
-                    // Agent count below FPS
-                    let alive = self.agents.iter().filter(|a| a.body.body.alive).count();
-                    text_items.push(TextItem {
-                        text: format!("Agents: {}/{}", alive, self.agents.len()),
-                        x: 0.62,
-                        y: 0.93,
-                        scale: 0.03,
-                        color: [0.0, 0.0, 0.0, 0.8],
-                    });
-
-                    // HUD bar labels
-                    let label_x = -0.98 + 0.36;
-                    let label_scale = 0.03;
-                    let label_color = [0.0, 0.0, 0.0, 0.9];
-                    let label_top = 0.97;
-                    let label_gap = 0.035;
-
-                    if let Some(agent) = self.agents.get(self.selected_agent_idx) {
-                        let energy = agent.body.body.internal.energy_signal();
-                        let integrity = agent.body.body.internal.integrity_signal();
-                        let pred_err = agent.cached_prediction_error;
-                        let explore = agent.cached_exploration_rate;
-
-                        text_items.push(TextItem {
-                            text: format!("Energy: {:.0}%", energy * 100.0),
-                            x: label_x,
-                            y: label_top,
-                            scale: label_scale,
-                            color: label_color,
-                        });
-                        text_items.push(TextItem {
-                            text: format!("Integrity: {:.0}%", integrity * 100.0),
-                            x: label_x,
-                            y: label_top - label_gap,
-                            scale: label_scale,
-                            color: label_color,
-                        });
-                        text_items.push(TextItem {
-                            text: format!("PredErr: {:.2}", pred_err),
-                            x: label_x,
-                            y: label_top - label_gap * 2.0,
-                            scale: label_scale,
-                            color: label_color,
-                        });
-                        text_items.push(TextItem {
-                            text: format!("Explore: {:.0}%", explore * 100.0),
-                            x: label_x,
-                            y: label_top - label_gap * 3.0,
-                            scale: label_scale,
-                            color: label_color,
-                        });
-
-                        // Info block below bars
-                        let info_y = label_top - label_gap * 4.5;
-                        let info_scale = 0.028;
-                        let info_color = [0.1, 0.1, 0.2, 0.9];
-                        text_items.push(TextItem {
-                            text: format!(
-                                "Agent {} | Gen {} | Deaths: {}",
-                                agent.id, agent.generation, agent.death_count
-                            ),
-                            x: -0.98,
-                            y: info_y,
-                            scale: info_scale,
-                            color: info_color,
-                        });
-                        text_items.push(TextItem {
-                            text: "Phase: GPU (TBD)".to_string(),
-                            x: -0.98,
-                            y: info_y - 0.035,
-                            scale: info_scale,
-                            color: info_color,
-                        });
-                    }
-
-                    if let Some(renderer) = &mut self.renderer {
-                        renderer.update_hud(&self.cached_hud_bars, &[]);
-                        renderer.update_text(&text_items);
-                    }
-
-                    self.hud_dirty = false;
-                }
-
-                // ── render (just draws — no geometry rebuilding) ──
-                let inst_buf = self.agent_instance_buffer.as_ref();
-                let inst_count = self.agent_instance_count;
-
-                let mut pending_evo_action = EvolutionAction::None;
-
-                if let Some(renderer) = &mut self.renderer {
-                    let t = self.terrain_gpu.as_ref();
-                    let f = self.food_gpu.as_ref();
-                    let h = self.heatmap_gpu.as_ref().filter(|g| g.num_indices > 0);
-                    let tr = self.trail_gpu.as_ref().filter(|g| g.num_indices > 0);
-                    let mk = self.marker_gpu.as_ref().filter(|g| g.num_indices > 0);
-                    let mut mesh_vec: Vec<&GpuMesh> = Vec::with_capacity(5);
-                    if let Some(t) = t {
-                        mesh_vec.push(t);
-                    }
-                    if let Some(f) = f {
-                        mesh_vec.push(f);
-                    }
-                    if let Some(h) = h {
-                        mesh_vec.push(h);
-                    }
-                    if let Some(tr) = tr {
-                        mesh_vec.push(tr);
-                    }
-                    if let Some(mk) = mk {
-                        mesh_vec.push(mk);
-                    }
-
-                    let vp = self.camera.view_projection_matrix();
-
-                    // ── Offscreen 3D → egui surface pipeline ──────────
-                    match renderer.begin_frame() {
-                        Ok(mut frame_ctx) => {
-                            // 1) Render 3D scene to offscreen viewport texture
-                            if self.render_3d {
-                                if let Some(egui) = &self.egui {
-                                    renderer.render_3d_offscreen(
-                                        &mesh_vec,
-                                        &vp,
-                                        inst_buf,
-                                        inst_count,
-                                        &mut frame_ctx.encoder,
-                                        &egui.viewport_color_view,
-                                        &egui.viewport_depth_view,
-                                    );
-                                }
-                            }
-
-                            // 2) Render egui UI to the surface (viewport texture embedded)
-                            if let (Some(egui), Some(window)) = (&mut self.egui, &self.window) {
-                                let screen = egui_wgpu::ScreenDescriptor {
-                                    size_in_pixels: [renderer.config.width, renderer.config.height],
-                                    pixels_per_point: window.scale_factor() as f32,
-                                };
-
-                                let fps = self.fps;
-                                let wall_time_secs = self.evo_snapshot.wall_time_secs;
-                                let ticks_per_sec = self.evo_snapshot.ticks_per_sec;
-                                let render_3d = self.render_3d;
-                                let speed_multiplier = self.speed_multiplier;
-                                let evo_state = self.evo_snapshot.state.clone();
-                                let viewport_tex_id = egui.viewport_texture_id;
-                                let ppp = window.scale_factor() as f32;
-                                let mut desired_vp = (0u32, 0u32);
-                                let selected_idx = self.selected_agent_idx;
-
-                                // Snapshot agent data for the UI closure (throttled to ~10 Hz)
-                                // Force rebuild when agents changed size or chart window changed.
-                                if self.cached_agent_snaps.len() != self.agents.len()
-                                    || self.chart_window != self.last_snap_chart_window
-                                {
-                                    self.snap_dirty = true;
-                                }
-                                let snap_rebuild_due = self.snap_dirty
-                                    && (self.paused
-                                        || self.last_snapshot_rebuild.elapsed()
-                                            >= REBUILD_THROTTLE);
-                                if snap_rebuild_due {
-                                    let snap_window = self.chart_window * 2;
-                                    self.cached_agent_snaps = self
-                                        .agents
-                                        .iter()
-                                        .map(|a| {
-                                            let tail =
-                                                |d: &std::collections::VecDeque<f32>| -> Vec<f32> {
-                                                    let skip = d.len().saturating_sub(snap_window);
-                                                    d.iter().skip(skip).copied().collect()
-                                                };
-                                            AgentSnapshot {
-                                                id: a.id,
-                                                generation: a.generation,
-                                                energy: a.body.body.internal.energy,
-                                                max_energy: a.body.body.internal.max_energy,
-                                                integrity: a.body.body.internal.integrity,
-                                                max_integrity: a.body.body.internal.max_integrity,
-                                                alive: a.body.body.alive,
-                                                deaths: a.death_count,
-                                                color: a.color,
-                                                longest_life: a.longest_life,
-                                                exploration_rate: a.cached_exploration_rate,
-                                                prediction_error: a.cached_prediction_error,
-                                                forward_weight_norm: 0.0, // GPU telemetry TBD
-                                                turn_weight_norm: 0.0,    // GPU telemetry TBD
-                                                prediction_error_history: tail(
-                                                    &a.prediction_error_history,
-                                                ),
-                                                exploration_rate_history: tail(
-                                                    &a.exploration_rate_history,
-                                                ),
-                                                energy_history: tail(&a.energy_history),
-                                                integrity_history: tail(&a.integrity_history),
-                                                gradient: a.cached_gradient,
-                                                urgency: a.cached_urgency,
-                                                food_consumed: a.food_consumed,
-                                                total_ticks_alive: a.total_ticks_alive,
-                                                motor_forward: a.cached_motor.forward,
-                                                motor_turn: a.cached_motor.turn,
-                                                phase: "GPU", // GPU telemetry TBD
-                                                vision_color: a.cached_frame.vision.color.clone(),
-                                                vision_width: a.cached_frame.vision.width,
-                                                vision_height: a.cached_frame.vision.height,
-                                                position: [
-                                                    a.body.body.position.x,
-                                                    a.body.body.position.y,
-                                                    a.body.body.position.z,
-                                                ],
-                                                yaw: a.body.yaw,
-                                                mean_attenuation: a.cached_mean_attenuation,
-                                                curiosity_bonus: a.cached_curiosity_bonus,
-                                                fatigue_factor: a.cached_fatigue_factor,
-                                                staleness: a.cached_staleness,
-                                                fatigue_history: tail(&a.fatigue_history),
-                                            }
-                                        })
-                                        .collect();
-                                    self.snap_dirty = false;
-                                    self.last_snap_chart_window = self.chart_window;
-                                    if !self.paused {
-                                        self.last_snapshot_rebuild = Instant::now();
-                                    }
-                                }
-                                let agent_snaps = self.cached_agent_snaps.as_slice();
-
-                                // Build evolution snapshot for the UI
-                                if let Some(gov) = &mut self.governor {
-                                    self.evo_snapshot.gen_tick = gov.gen_tick;
-                                    self.evo_snapshot.generation = gov.generation;
-                                    self.evo_snapshot.tree_nodes = gov.tree_nodes();
-                                    self.evo_snapshot.current_node_id = gov.current_node_id;
-                                    self.evo_snapshot
-                                        .fitness_history
-                                        .clone_from(gov.fitness_history_by_island());
-                                    self.evo_snapshot.best_fitness = gov.best_score();
-                                }
-                                let wall = self.evo_wall_accumulated
-                                    + self
-                                        .evo_wall_segment_start
-                                        .map(|s| s.elapsed().as_secs_f64())
-                                        .unwrap_or(0.0);
-                                self.evo_snapshot.wall_time_secs = wall;
-                                let tps_elapsed = self.tps_last_reset.elapsed().as_secs_f64();
-                                if tps_elapsed >= 1.0 {
-                                    self.tps_display = self.tps_tick_count as f64 / tps_elapsed;
-                                    self.tps_tick_count = 0;
-                                    self.tps_last_reset = Instant::now();
-                                }
-                                self.evo_snapshot.ticks_per_sec = self.tps_display;
-                                // Move snapshot out so we can pass &mut to the closure
-                                let mut evo_snap = std::mem::take(&mut self.evo_snapshot);
-                                let gen_tick = evo_snap.gen_tick;
-                                let tick_budget = evo_snap.tick_budget;
-                                let evo_generation = evo_snap.generation;
-                                let best_fitness = evo_snap.best_fitness;
-                                let mut evo_action = EvolutionAction::None;
-
-                                // Build world snapshot for mini-map
-                                if let Some(world) = &self.world {
-                                    let gpu_food = self
-                                        .gpu_kernel
-                                        .as_ref()
-                                        .and_then(|mk| mk.cached_food_state());
-                                    if let Some(food) = gpu_food {
-                                        self.world_snapshot.food_positions = food
-                                            .chunks_exact(FOOD_STATE_STRIDE)
-                                            .filter(|c| c[F_RESPAWN_TIMER] <= 0.0)
-                                            .map(|c| [c[F_POS_X], c[F_POS_Z]])
-                                            .collect();
-                                    } else {
-                                        self.world_snapshot.food_positions = world
-                                            .food_items
-                                            .iter()
-                                            .filter(|f| !f.consumed)
-                                            .map(|f| [f.position.x, f.position.z])
-                                            .collect();
-                                    }
-                                    self.world_snapshot.world_size = world.config.world_size;
-                                }
-                                let mut world_snap = std::mem::take(&mut self.world_snapshot);
-                                // Pre-build biome image pixels if texture not yet created
-                                let mut biome_image = if world_snap.biome_texture.is_none() {
-                                    self.world.as_ref().map(|world| {
-                                        use xagent_sandbox::world::biome::BiomeType;
-                                        let res = 256usize;
-                                        let ws = world.config.world_size;
-                                        let half = ws / 2.0;
-                                        let cell = ws / res as f32;
-                                        let mut pixels = Vec::with_capacity(res * res);
-                                        for row in 0..res {
-                                            let z = -half + (row as f32 + 0.5) * cell;
-                                            for col in 0..res {
-                                                let x = -half + (col as f32 + 0.5) * cell;
-                                                let biome = world.biome_map.biome_at(x, z);
-                                                let c = match biome {
-                                                    BiomeType::FoodRich => {
-                                                        egui::Color32::from_rgb(25, 70, 20)
-                                                    }
-                                                    BiomeType::Barren => {
-                                                        egui::Color32::from_rgb(60, 50, 30)
-                                                    }
-                                                    BiomeType::Danger => {
-                                                        egui::Color32::from_rgb(80, 25, 15)
-                                                    }
-                                                };
-                                                pixels.push(c);
-                                            }
-                                        }
-                                        egui::ColorImage {
-                                            size: [res, res],
-                                            pixels,
-                                        }
-                                    })
-                                } else {
-                                    None
-                                };
-
-                                let console_lines: Vec<&str> =
-                                    self.console_log.iter().map(|s| s.as_str()).collect();
-
-                                let mut clicked_agent_idx: Option<usize> = None;
-                                let mut open_agent_tab: Option<u32> = None;
-                                let mut vp_hovered = false;
-                                let mut chart_win = self.chart_window;
-                                let mut sort_mode = self.sort_mode;
-                                let dock_state = &mut self.dock_state;
-                                let replay_state = &mut self.replay_state;
-                                let last_recording = self.last_recording.as_ref();
-                                let orbit_mode = &mut self.orbit_mode;
-
-                                egui.render(
-                                    window,
-                                    &renderer.device,
-                                    &renderer.queue,
-                                    &mut frame_ctx.encoder,
-                                    &frame_ctx.view,
-                                    screen,
-                                    |ctx| {
-                                        // Build biome texture once (requires ctx)
-                                        if let Some(image) = biome_image.take() {
-                                            world_snap.biome_texture = Some(
-                                                ctx.load_texture("biome_map", image, egui::TextureOptions::NEAREST)
-                                            );
-                                        }
-
-                                        // ── Top bar ──────────────────────────
-                                        egui::TopBottomPanel::top("top_bar").show(ctx, |ui| {
-                                            ui.horizontal(|ui| {
-                                                ui.label(
-                                                    egui::RichText::new("xagent")
-                                                        .strong()
-                                                        .color(egui::Color32::from_rgb(120, 200, 255)),
-                                                );
-                                                ui.separator();
-                                                ui.label(format!("FPS: {:.0}", fps));
-                                                ui.separator();
-                                                ui.label(format!("Agents: {}", agent_snaps.len()));
-                                                ui.separator();
-                                                match &evo_state {
-                                                    EvolutionState::Idle => {
-                                                        ui.label(
-                                                            egui::RichText::new("⏹ IDLE")
-                                                                .color(egui::Color32::GRAY),
-                                                        );
-                                                    }
-                                                    EvolutionState::HasSession { generation } => {
-                                                        ui.label(
-                                                            egui::RichText::new(format!("💾 SESSION @ Gen {}", generation))
-                                                                .color(egui::Color32::from_rgb(100, 180, 255)),
-                                                        );
-                                                    }
-                                                    EvolutionState::Running => {
-                                                        ui.label(
-                                                            egui::RichText::new("▶ RUNNING")
-                                                                .color(egui::Color32::from_rgb(50, 200, 80)),
-                                                        );
-                                                    }
-                                                    EvolutionState::Paused => {
-                                                        ui.label(
-                                                            egui::RichText::new("⏸ PAUSED")
-                                                                .color(egui::Color32::YELLOW),
-                                                        );
-                                                    }
-                                                }
-                                                if matches!(&evo_state, EvolutionState::Running | EvolutionState::Paused) {
-                                                    ui.separator();
-                                                    let hours = (wall_time_secs / 3600.0) as u64;
-                                                    let mins = ((wall_time_secs % 3600.0) / 60.0) as u64;
-                                                    let secs = (wall_time_secs % 60.0) as u64;
-                                                    ui.label(format!("{}h {:02}m {:02}s", hours, mins, secs));
-                                                    ui.separator();
-                                                    let speed_label = speed_label(speed_multiplier);
-                                                    ui.label(
-                                                        egui::RichText::new(format!("⏩ {}", speed_label))
-                                                            .color(if speed_multiplier > 1 {
-                                                                egui::Color32::from_rgb(255, 200, 50)
-                                                            } else {
-                                                                egui::Color32::GRAY
-                                                            }),
-                                                    );
-                                                    ui.separator();
-                                                    ui.label(format!("{:.0} ticks/s", ticks_per_sec));
-                                                }
-                                                if matches!(&evo_state, EvolutionState::Running | EvolutionState::Paused) && best_fitness >= 0.0 {
-                                                    ui.separator();
-                                                    ui.label(
-                                                        egui::RichText::new(format!("Best: {:.4}", best_fitness))
-                                                            .color(egui::Color32::from_rgb(50, 200, 80)),
-                                                    );
-                                                }
-                                                // Generation progress bar (compact, in toolbar)
-                                                if matches!(&evo_state, EvolutionState::Running | EvolutionState::Paused) && tick_budget > 0 {
-                                                    ui.separator();
-                                                    let progress = gen_tick as f32 / tick_budget as f32;
-                                                    ui.add(
-                                                        egui::ProgressBar::new(progress)
-                                                            .text(format!("Gen {} — {:.0}%", evo_generation, progress * 100.0))
-                                                            .desired_width(160.0)
-                                                            .animate(matches!(&evo_state, EvolutionState::Running)),
-                                                    );
-                                                }
-                                                if !render_3d {
-                                                    ui.separator();
-                                                    ui.label(
-                                                        egui::RichText::new("⚡ FAST (G)")
-                                                            .color(egui::Color32::from_rgb(255, 160, 50)),
-                                                    );
-                                                }
-                                                // ── Right-aligned controls ──
-                                                ui.with_layout(egui::Layout::right_to_left(egui::Align::Center), |ui| {
-                                                    if matches!(&evo_state, EvolutionState::Running | EvolutionState::Paused) {
-                                                        if ui.add(egui::Button::new(
-                                                            egui::RichText::new("🗑 Reset")
-                                                                .color(egui::Color32::from_rgb(220, 80, 80)),
-                                                        )).clicked() {
-                                                            evo_action = EvolutionAction::Reset;
-                                                        }
-                                                        if matches!(&evo_state, EvolutionState::Running) {
-                                                            if ui.button("⏸ Pause").clicked() {
-                                                                evo_action = EvolutionAction::Pause;
-                                                            }
-                                                        } else {
-                                                            if ui.button("▶ Resume").clicked() {
-                                                                evo_action = EvolutionAction::Unpause;
-                                                            }
-                                                        }
-                                                    }
-                                                });
-                                            });
-                                        });
-
-                                        // ── Bottom console ───────────────────
-                                        egui::TopBottomPanel::bottom("console")
-                                            .resizable(true)
-                                            .default_height(120.0)
-                                            .show(ctx, |ui| {
-                                                ui.label(
-                                                    egui::RichText::new("Console")
-                                                        .small()
-                                                        .color(egui::Color32::GRAY),
-                                                );
-                                                ui.separator();
-                                                let scroll_width = ui.available_width();
-                                                egui::ScrollArea::vertical()
-                                                    .stick_to_bottom(true)
-                                                    .min_scrolled_width(scroll_width)
-                                                    .show(ui, |ui| {
-                                                        ui.set_min_width(scroll_width);
-                                                        for line in &console_lines {
-                                                            let color = if line.contains("ERROR") || line.contains("Failed to") {
-                                                                egui::Color32::from_rgb(255, 100, 100)
-                                                            } else if line.contains("best") || line.contains("Beat parent") {
-                                                                egui::Color32::from_rgb(80, 220, 80)
-                                                            } else if line.contains("Failed") || line.contains("exhausted") || line.contains("backtracking") {
-                                                                egui::Color32::from_rgb(255, 140, 80)
-                                                            } else if line.contains("Migration") {
-                                                                egui::Color32::from_rgb(100, 180, 255)
-                                                            } else if line.contains("Momentum") {
-                                                                egui::Color32::from_rgb(180, 160, 255)
-                                                            } else {
-                                                                egui::Color32::LIGHT_GRAY
-                                                            };
-                                                            ui.label(
-                                                                egui::RichText::new(*line)
-                                                                    .monospace()
-                                                                    .size(11.0)
-                                                                    .color(color),
-                                                            );
-                                                        }
-                                                    });
-                                            });
-
-                                        // ── Left sidebar: agent list ─────────
-                                        egui::SidePanel::left("agent_list")
-                                            .resizable(true)
-                                            .default_width(200.0)
-                                            .show(ctx, |ui| {
-                                                ui.label(
-                                                    egui::RichText::new("Agents")
-                                                        .strong()
-                                                        .size(14.0),
-                                                );
-                                                ui.separator();
-                                                ui.horizontal(|ui| {
-                                                    ui.label(egui::RichText::new("Sort:").small().color(egui::Color32::GRAY));
-                                                    egui::ComboBox::from_id_salt("agent_sort")
-                                                        .selected_text(sort_mode.label())
-                                                        .width(90.0)
-                                                        .show_ui(ui, |ui| {
-                                                            for mode in SortMode::ALL {
-                                                                ui.selectable_value(&mut sort_mode, mode, mode.label());
-                                                            }
-                                                        });
-                                                });
-                                                ui.add_space(2.0);
-                                                egui::ScrollArea::vertical().show(ui, |ui| {
-                                                    let mut sorted_indices: Vec<usize> = (0..agent_snaps.len()).collect();
-                                                    match sort_mode {
-                                                        SortMode::Id => {} // already sorted by id
-                                                        SortMode::Energy => {
-                                                            sorted_indices.sort_by(|&a, &b| {
-                                                                let ea = agent_snaps[a].energy / agent_snaps[a].max_energy.max(0.001);
-                                                                let eb = agent_snaps[b].energy / agent_snaps[b].max_energy.max(0.001);
-                                                                eb.partial_cmp(&ea).unwrap_or(std::cmp::Ordering::Equal)
-                                                            });
-                                                        }
-                                                        SortMode::Integrity => {
-                                                            sorted_indices.sort_by(|&a, &b| {
-                                                                let ia = agent_snaps[a].integrity / agent_snaps[a].max_integrity.max(0.001);
-                                                                let ib = agent_snaps[b].integrity / agent_snaps[b].max_integrity.max(0.001);
-                                                                ib.partial_cmp(&ia).unwrap_or(std::cmp::Ordering::Equal)
-                                                            });
-                                                        }
-                                                        SortMode::Deaths => {
-                                                            sorted_indices.sort_by(|&a, &b| {
-                                                                agent_snaps[a].deaths.cmp(&agent_snaps[b].deaths)
-                                                            });
-                                                        }
-                                                        SortMode::LongestLife => {
-                                                            sorted_indices.sort_by(|&a, &b| {
-                                                                agent_snaps[b].longest_life.cmp(&agent_snaps[a].longest_life)
-                                                            });
-                                                        }
-                                                        SortMode::PredictionError => {
-                                                            sorted_indices.sort_by(|&a, &b| {
-                                                                agent_snaps[a].prediction_error.partial_cmp(&agent_snaps[b].prediction_error)
-                                                                    .unwrap_or(std::cmp::Ordering::Equal)
-                                                            });
-                                                        }
-                                                        SortMode::Fitness => {
-                                                            sorted_indices.sort_by(|&a, &b| {
-                                                                let fa = agent_snaps[a].food_consumed as f64
-                                                                    + agent_snaps[a].total_ticks_alive as f64 * 0.001;
-                                                                let fb = agent_snaps[b].food_consumed as f64
-                                                                    + agent_snaps[b].total_ticks_alive as f64 * 0.001;
-                                                                fb.partial_cmp(&fa).unwrap_or(std::cmp::Ordering::Equal)
-                                                            });
-                                                        }
-                                                    }
-                                                    for &idx in &sorted_indices {
-                                                        let snap = &agent_snaps[idx];
-                                                        let is_selected = idx == selected_idx;
-                                                        let color = egui::Color32::from_rgb(
-                                                            (snap.color[0] * 255.0) as u8,
-                                                            (snap.color[1] * 255.0) as u8,
-                                                            (snap.color[2] * 255.0) as u8,
-                                                        );
-                                                        let frame = if is_selected {
-                                                            egui::Frame::NONE
-                                                                .fill(egui::Color32::from_rgba_premultiplied(60, 60, 80, 255))
-                                                                .inner_margin(4.0)
-                                                                .corner_radius(3.0)
-                                                        } else {
-                                                            egui::Frame::NONE
-                                                                .inner_margin(4.0)
-                                                        };
-                                                        let response = frame.show(ui, |ui| {
-                                                            ui.horizontal(|ui| {
-                                                                let (rect, _) = ui.allocate_exact_size(
-                                                                    egui::vec2(10.0, 10.0),
-                                                                    egui::Sense::hover(),
-                                                                );
-                                                                ui.painter().circle_filled(
-                                                                    rect.center(),
-                                                                    5.0,
-                                                                    color,
-                                                                );
-                                                                let status = if !snap.alive { "💀" } else { "" };
-                                                                ui.label(format!(
-                                                                    "Agent {} (g{}) {}",
-                                                                    snap.id, snap.generation, status
-                                                                ));
-                                                            });
-                                                            ui.horizontal(|ui| {
-                                                                ui.label(
-                                                                    egui::RichText::new(snap.phase)
-                                                                        .small()
-                                                                        .color(match snap.phase {
-                                                                            "ADAPTED" => egui::Color32::from_rgb(80, 200, 80),
-                                                                            "LEARNING" => egui::Color32::from_rgb(200, 200, 80),
-                                                                            "EXPLORING" => egui::Color32::from_rgb(200, 140, 60),
-                                                                            _ => egui::Color32::from_rgb(150, 150, 150),
-                                                                        }),
-                                                                );
-                                                                ui.label(
-                                                                    egui::RichText::new(format!(
-                                                                        "| D:{} F:{}",
-                                                                        snap.deaths, snap.food_consumed
-                                                                    ))
-                                                                    .small()
-                                                                    .color(egui::Color32::GRAY),
-                                                                );
-                                                            });
-                                                        });
-                                                        let resp = response.response.interact(egui::Sense::click());
-                                                        if resp.clicked() {
-                                                            clicked_agent_idx = Some(idx);
-                                                        }
-                                                        if resp.double_clicked() {
-                                                            open_agent_tab = Some(snap.id);
-                                                        }
-                                                        ui.add_space(2.0);
-                                                    }
-                                                });
-                                            });
-
-                                        // ── Central dock area (tabs) ─────────
-                                        egui::CentralPanel::default()
-                                            .frame(egui::Frame::NONE)
-                                            .show(ctx, |ui| {
-                                                let mut tab_ctx = TabContext {
-                                                    viewport_tex_id,
-                                                    ppp,
-                                                    desired_vp: &mut desired_vp,
-                                                    viewport_hovered: &mut vp_hovered,
-                                                    chart_window: &mut chart_win,
-                                                    agents: agent_snaps,
-                                                    evolution: &mut evo_snap,
-                                                    evolution_action: &mut evo_action,
-                                                    world: &world_snap,
-                                                    replay: replay_state,
-                                                    recording: last_recording,
-                                                    orbit_mode,
-                                                };
-                                                egui_dock::DockArea::new(dock_state)
-                                                    .style(egui_dock::Style::from_egui(ui.style().as_ref()))
-                                                    .show_inside(ui, &mut tab_ctx);
-                                            });
-                                    },
-                                );
-
-                                self.viewport_hovered = vp_hovered;
-                                self.chart_window = chart_win;
-                                self.sort_mode = sort_mode;
-
-                                // Restore evolution snapshot (may have been mutated by UI)
-                                self.evo_snapshot = evo_snap;
-                                // Restore world snapshot (biome texture may have been created)
-                                self.world_snapshot = world_snap;
-
-                                // Defer evolution action to after the renderer borrow
-                                pending_evo_action = evo_action;
-
-                                // Handle agent selection from sidebar click
-                                if let Some(idx) = clicked_agent_idx {
-                                    self.selected_agent_idx = idx;
-                                }
-
-                                // Handle double-click → open agent detail tab
-                                if let Some(agent_id) = open_agent_tab {
-                                    let tab = Tab::AgentDetail(agent_id);
-                                    // Check if tab already exists
-                                    let already_open =
-                                        self.dock_state.iter_all_tabs().any(|(_, t)| *t == tab);
-                                    if !already_open {
-                                        self.dock_state.push_to_focused_leaf(tab);
-                                    }
-                                }
-
-                                // Resize offscreen textures if the panel changed size (takes effect next frame)
-                                if desired_vp.0 > 0 && desired_vp.1 > 0 {
-                                    egui.resize_viewport(
-                                        &renderer.device,
-                                        desired_vp.0,
-                                        desired_vp.1,
-                                    );
-                                    // Update camera aspect to match viewport
-                                    self.camera.aspect =
-                                        desired_vp.0 as f32 / desired_vp.1.max(1) as f32;
-                                }
-                            }
-
-                            renderer.finish_frame(frame_ctx);
-                        }
-                        Err(wgpu::SurfaceError::Lost) => {
-                            let w = renderer.config.width;
-                            let h = renderer.config.height;
-                            renderer.resize(w, h);
-                        }
-                        Err(wgpu::SurfaceError::OutOfMemory) => {
-                            log::error!("Out of memory!");
-                            self.print_session_summary();
-                            event_loop.exit();
-                        }
-                        Err(e) => {
-                            log::warn!("Surface error: {:?}", e);
-                        }
-                    }
-                }
+                // ── render (3D offscreen + egui chrome) ──
+                let pending_evo_action = self.render_frame(event_loop);
 
                 // Poll GPU kernel device after rendering so map_async
                 // callbacks can complete before the next frame.
